fix: make initial page visibility configurable (#26)
Previously, the page visibility state was initialized from the `document.hidden` state. When server-rendering, `document` is `undefined` and throws an error. Instead, we assume the page is visible on load by default, and allow a custom initial value to be passed in so client-rendered apps can keep using `document.hidden`.